refactor(middleware): extract session/proof + _kyaos protocol sub-factories
Move handshake establishment, ensureSession, wrapWithProof and attachOutcomeProof (the activeSessionId owner) into with-kya-os.session.ts, and the _kyaos action router + identity responder + tool definitions into with-kya-os.protocol.ts. The orchestrator (createKyaOsMiddleware) is now a ~180-line composition root that constructs deps and wires the sub-factories. No behaviour change; 1607 tests green. Phase 2d.
